backend/requirements.txt
# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9

# Configurazione
pydantic-settings>=2.5.0

# HTTP client asincrono
httpx>=0.27.0

# Image processing
Pillow>=10.0.0
numpy>=1.26.0

# PDF processing
PyMuPDF>=1.24.0

# AI providers - Livello 1 (a pagamento)
anthropic>=0.34.0

# AI providers - Livello 2 (gratuito, opzionale)
google-genai>=0.21.0

# OCR fallback
pytesseract>=0.3.13

# QR Code e DataMatrix nativi (richiede libzbar0 e libdmtx-dev come dipendenze di sistema)
pyzbar>=0.1.9
pylibdmtx>=0.1.10

# Fuzzy matching per catalogo tipi macchina
rapidfuzz>=3.0.0

# Rate limiting
slowapi>=0.1.9

# Web scraping fallback
beautifulsoup4>=4.12.0
ddgs>=0.1.0

# Database (Supabase via connessione diretta Postgres)
psycopg2-binary>=2.9.0

# RAG corpus normativo — ChromaDB + sentence-transformers
# chromadb e sentence-transformers vengono caricati lazy: se non installati
# il sistema degrada silenziosamente (solo dizionario D.Lgs 81/08 attivo).
chromadb==0.5.0
sentence-transformers==3.0.0
langchain-text-splitters==0.2.0

# Crawler notturno — job scheduler asincrono
apscheduler>=3.10.0